Have the application obj require the library dependencies
Previously the binary required the library which means the main object
files maybe compiled before libraries are generated. By having the
application 'obj' require the library we know the libraries will be
generated before compiling any main objects.

In this trivial code this isn't necessary I think it's better because
if a side effect of generating the library caused additional headers
or code to be created that work will be completed before the application
is compiled.